refactor: organize mobile routers folder
