+++ packages/mobile/lib/routers/index.ts
@@ -1,9 +1,5 @@
 export * from './actions'
 export * from './enums'
 export * from './interfaces'
+export * from './routers'
 export * from './utils'
-
-export * from './app-router'
-export * from './dashboard-router'
-export * from './onboarding-router'
-export * from './subrouters'
