+++ packages/mobile/lib/routers/routers/app-router.ts
@@ -4,7 +4,7 @@ import { profiles } from '@core/profile'
 import { IRouterEvent, Router } from '@core/router'
 
 import { AppRoute, LoginRoute } from './enums'
-import { loginRoute } from './subrouters'
+import { loginRoute } from '.'
 
 export const appRoute = writable<AppRoute>(null)
 export const appRouter = writable<AppRouter>(null)
