+++ packages/mobile/lib/routers/routers/dashboard/index.ts
@@ -1,8 +1,4 @@
-export * from './onboarding'
-
 export * from './account-switcher-router'
 export * from './activity-router'
-export * from './login-router'
-export * from './onboarding'
 export * from './send-router'
 export * from './token-router'
