+++ packages/mobile/lib/routers/routers/index.ts
@@ -1,0 +1,7 @@
+export * from './app-router'
+export * from './dashboard-router'
+export * from './onboarding-router'
+export * from './login-router'
+
+export * from './dashboard'
+export * from './onboarding'
