+++ packages/mobile/lib/routers/routers/onboarding/storage-protection-setup-router.ts
@@ -3,7 +3,7 @@ import { get, writable } from 'svelte/store'
 import { onboardingProfile, ProfileProtectionType } from '@contexts/onboarding'
 import { Subrouter } from '@core/router'
 
-import { onboardingRouter } from '../../onboarding-router'
+import { onboardingRouter } from '../onboarding-router'
 import { StorageProtectionSetupRoute } from '../../enums'
 
 export const storageProtectionSetupRoute = writable<StorageProtectionSetupRoute>(null)
