+++ packages/mobile/lib/routers/utils/getAppRouter.ts
@@ -1,7 +1,7 @@
 import { get } from 'svelte/store'
 
 import { IRouter } from '@core/router/interfaces'
-import { appRouter } from '../app-router'
+import { appRouter } from '../routers'
 
 export function getAppRouter(): IRouter {
     return get(appRouter)
